refactor: create 메서드 대신 builder 사용
