+++ src/main/java/nora/movlog/domain/movie/Actor.java
@@ -2,13 +2,12 @@ package nora.movlog.domain.movie;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
+@Builder
+@Getter
 @Entity
 public class Actor {
     /* 데이터 관리 및 조회용 */
@@ -17,8 +16,4 @@ public class Actor {
 
     /* 배우 데이터 */
     private String name;    // 영어 이름
-
-    public static Actor create(String id, String name) {
-        return new Actor(id, name);
-    }
 }
