+++ src/main/java/nora/movlog/domain/movie/Director.java
@@ -2,13 +2,12 @@ package nora.movlog.domain.movie;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
+@Builder
+@Getter
 @Entity
 public class Director {
     /* 데이터 관리 및 조회용 */
@@ -17,8 +16,4 @@ public class Director {
 
     /* 감독 데이터 */
     private String name;    // 한국 이름
-
-    public static Director create(String id, String name) {
-        return new Director(id, name);
-    }
 }
