+++ src/main/java/nora/movlog/domain/movie/Genre.java
@@ -2,13 +2,12 @@ package nora.movlog.domain.movie;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
+@Builder
+@Getter
 @Entity
 public class Genre {
     /* 데이터 관리 및 조회용 */
@@ -17,8 +16,4 @@ public class Genre {
 
     /* 장르 데이터 */
     private String name;       // 한국 이름
-
-    public static Genre create(Integer id, String name) {
-        return new Genre(id, name);
-    }
 }
