+++ src/main/java/nora/movlog/domain/movie/Nation.java
@@ -2,20 +2,15 @@ package nora.movlog.domain.movie;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
+@Builder
+@Getter
 @Entity
 public class Nation {
     /* 데이터 관리 및 조회용 */
     @Id
     private String id;      // DB ID - iso-3166 code
-
-    public static Nation create(String id) {
-        return new Nation(id);
-    }
 }
