+++ src/main/java/nora/movlog/repository/movie/MovieKobisApiRepository.java
@@ -73,6 +73,4 @@ public class MovieKobisApiRepository {
 
         return new Movie();
     }
-
-
 }
