+++ src/main/java/nora/movlog/service/movie/NationService.java
@@ -24,7 +24,9 @@ public class NationService {
 
             if (nation.isPresent()) nations.add(nation.get());
             else {
-                nationRepository.save(Nation.create(id));
+                nationRepository.save(Nation.builder()
+                        .id(id)
+                        .build());
                 nations.add(nationRepository.findById(id).get());
             }
         }
